requirements.txt
flask
argon2-cffi
